+++ ConsoleApplication2/tools_vssd.h
@@ -14,7 +14,7 @@ namespace sjh {
 
 			static void				PushString(std::wstring Str, std::vector<wchar_t> &Byte);
 			static void				PushWcharVector(std::vector<wchar_t> &StringArray, std::vector<wchar_t> &Byte);
-			static size_t			PushLengthValue(unsigned int Uint4, std::vector<wchar_t> &Byte);
+			static size_t			PushLengthValue(size_t Uint4, std::vector<wchar_t> &Byte);
 			static void				PushToRealFile(std::wstring JumpTo, std::vector<wchar_t> aSerial);
 
 			static std::wstring		GetString(const std::vector<wchar_t>& ByteVssd, int &Pos);
@@ -22,9 +22,9 @@ namespace sjh {
 			static unsigned int		GetLengthValue(const std::vector<wchar_t>& ByteVssd, int &Pos);
 			static void				GetFromRealfile(std::wstring GetFrom, std::vector<wchar_t> aSerial);
 			static std::wstring		GetTimeWString(time_t tm);
-			static std::wstring		GetSizeWString(unsigned int Size);
+			static std::wstring		GetSizeWString(size_t Size);
 			static std::string		GetTimeString(time_t tm);
-			static std::string		GetSizeString(unsigned int Size);
+			static std::string		GetSizeString(size_t Size);
 		};
 		 
 	}
